# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is Swagger UI environment generated for the TM Forum Intent Management\
    \ API specification"
  version: "4.0.0"
  title: "Intent"
host: "serverRoot:8080"
basePath: "/tmf-api/intent/v4/"
...
schemes:
- "https"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
  /intent:
    get:
      tags:
      - "intent"
      summary: "List or find Intent objects"
      description: "This operation list or find Intent entities"
      operationId: "listIntent"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Intent"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Intent"
    post:
      tags:
      - "intent"
      summary: "Creates a Intent"
      description: "This operation creates a Intent entity."
      operationId: "createIntent"
      parameters:
      - in: "body"
        name: "intent"
        description: "The Intent to be created"
        required: true
        schema:
          $ref: "#/definitions/Intent_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/Intent"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Intent"
  /intent/{id}:
    get:
      tags:
      - "intent"
      summary: "Retrieves a Intent by ID"
      description: "This operation retrieves a Intent entity. Attribute selection\
        \ is enabled for all first level attributes."
      operationId: "retrieveIntent"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Intent"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Intent"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Intent"
    delete:
      tags:
      - "intent"
      summary: "Deletes a Intent"
      description: "This operation deletes a Intent entity."
      operationId: "deleteIntent"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Intent"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Intent"
    patch:
      tags:
      - "intent"
      summary: "Updates partially a Intent"
      description: "This operation updates partially a Intent entity."
      operationId: "patchIntent"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Intent"
        required: true
        type: "string"
      - in: "body"
        name: "intent"
        description: "The Intent to be updated"
        required: true
        schema:
          $ref: "#/definitions/Intent_Update"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/Intent"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Intent"
...
definitions:
...
  Expression:
    type: "object"
    properties:
      expressionLanguage:
        type: "string"
        description: "The serialized Expression language of the intent serialized."
      iri:
        type: "string"
        description: "Internationalized Resource Identifier of the intent Expression"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class Extensible name"
    description: "Expression is the ontology-encoded form of the Intent"
    example:
      iri: "iri"
      '@baseType': "@baseType"
      '@type': "@type"
      expressionLanguage: "expressionLanguage"
      '@schemaLocation': "http://example.com/aeiou"
...
  Intent:
    type: "object"
    properties:
      id:
        type: "string"
        description: "unique identifier"
      href:
        type: "string"
        format: "uri"
        description: "Hyperlink reference"
      creationDate:
        type: "string"
        format: "date-time"
        description: "Date and time of the creation of this Intent"
      description:
        type: "string"
        description: "The description of the intent."
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update of this Intent"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status of this Intent"
      name:
        type: "string"
        description: "The name of the intent."
      statusChangeDate:
        type: "string"
        format: "date-time"
        description: "A date time( DateTime). The date that the entity status changed\
          \ to the current one"
      version:
        type: "string"
        description: "A field that identifies the specific version of an instance\
          \ of an entity. If not specified version is defaulted to 1.0"
      expression:
        description: "An Intent contains a set of expectations expressed  (expression)\
          \ in statements using a particular expressionLanguage. This attribute captures\
          \ the intent expressed by the TM Forum Intent Ontology (TIO) as is stored\
          \ as a string. The Expression is the formal statement of the intent and\
          \ contains all the information that describes the complete intent."
        $ref: "#/definitions/Expression"
      validFor:
        description: "The period for which this intent is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing Intent, this defines the sub-class Extensible\
          \ name"
        enum:
        - "Intent"
        - "ProbeIntent"
        - "JudgeIntent"
        - "BestIntent"
    description: "An Intent instance is the formal description of all expectations\
      \ including requirements, goals, and constraints given to a technical system"
    example:
      expression:
        iri: "iri"
        '@baseType': "@baseType"
        '@type': "@type"
        expressionLanguage: "expressionLanguage"
        '@schemaLocation': "http://example.com/aeiou"
      lifecycleStatus: "lifecycleStatus"
      validFor:
        startDateTime: "1985-04-12T23:20:50.52Z"
        endDateTime: "1985-04-12T23:20:50.52Z"
      '@type': "Intent"
      description: "description"
      creationDate: "2000-01-23T04:56:07.000+00:00"
      version: "version"
      statusChangeDate: "2000-01-23T04:56:07.000+00:00"
      '@baseType': "@baseType"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      id: "id"
      href: "http://example.com/aeiou"
      '@schemaLocation': "http://example.com/aeiou"
  Intent_Create:
    type: "object"
    required:
    - "expression"
    - "name"
    properties:
      creationDate:
        type: "string"
        format: "date-time"
        description: "Date and time of the creation of this Intent"
      description:
        type: "string"
        description: "The description of the intent."
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update of this Intent"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status of this Intent"
      name:
        type: "string"
        description: "The name of the intent."
      statusChangeDate:
        type: "string"
        format: "date-time"
        description: "A date time( DateTime). The date that the entity status changed\
          \ to the current one"
      version:
        type: "string"
        description: "A field that identifies the specific version of an instance\
          \ of an entity. If not specified version is defaulted to 1.0"
      expression:
        description: "An Intent contains a set of expectations expressed  (expression)\
          \ in statements using a particular expressionLanguage. This attribute captures\
          \ the intent expressed by the TM Forum Intent Ontology (TIO) as is stored\
          \ as a string. The Expression is the formal statement of the intent and\
          \ contains all the information that describes the complete intent."
        $ref: "#/definitions/Expression"
      validFor:
        description: "The period for which this intent is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class Extensible name"
    description: "An Intent instance is the formal description of all expectations\
      \ including requirements, goals, and constraints given to a technical system\n\
      Skipped properties: id,href"
    example:
      statusChangeDate: "2000-01-23T04:56:07.000+00:00"
      expression:
        iri: "iri"
        '@baseType': "@baseType"
        '@type': "@type"
        expressionLanguage: "expressionLanguage"
        '@schemaLocation': "http://example.com/aeiou"
      lifecycleStatus: "lifecycleStatus"
      '@baseType': "@baseType"
      validFor:
        startDateTime: "1985-04-12T23:20:50.52Z"
        endDateTime: "1985-04-12T23:20:50.52Z"
      '@type': "@type"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      description: "description"
      creationDate: "2000-01-23T04:56:07.000+00:00"
      '@schemaLocation': "http://example.com/aeiou"
      version: "version"
  Intent_Update:
    type: "object"
    properties:
      creationDate:
        type: "string"
        format: "date-time"
        description: "Date and time of the creation of this Intent"
      description:
        type: "string"
        description: "The description of the intent."
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update of this Intent"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status of this Intent"
      name:
        type: "string"
        description: "The name of the intent."
      statusChangeDate:
        type: "string"
        format: "date-time"
        description: "A date time( DateTime). The date that the entity status changed\
          \ to the current one"
      version:
        type: "string"
        description: "A field that identifies the specific version of an instance\
          \ of an entity. If not specified version is defaulted to 1.0"
      expression:
        description: "An Intent contains a set of expectations expressed  (expression)\
          \ in statements using a particular expressionLanguage. This attribute captures\
          \ the intent expressed by the TM Forum Intent Ontology (TIO) as is stored\
          \ as a string. The Expression is the formal statement of the intent and\
          \ contains all the information that describes the complete intent."
        $ref: "#/definitions/Expression"
      validFor:
        description: "The period for which this intent is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class Extensible name"
    description: "An Intent instance is the formal description of all expectations\
      \ including requirements, goals, and constraints given to a technical system\n\
      Skipped properties: id,href"
    example:
      statusChangeDate: "2000-01-23T04:56:07.000+00:00"
      expression:
        iri: "iri"
        '@baseType': "@baseType"
        '@type': "@type"
        expressionLanguage: "expressionLanguage"
        '@schemaLocation': "http://example.com/aeiou"
      lifecycleStatus: "lifecycleStatus"
      '@baseType': "@baseType"
      validFor:
        startDateTime: "1985-04-12T23:20:50.52Z"
        endDateTime: "1985-04-12T23:20:50.52Z"
      '@type': "@type"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      description: "description"
      creationDate: "2000-01-23T04:56:07.000+00:00"
      '@schemaLocation': "http://example.com/aeiou"
      version: "version"
...
  TimePeriod:
    type: "object"
    properties:
      endDateTime:
        type: "string"
        format: "date-time"
        example: "1985-04-12T23:20:50.52Z"
        description: "End of the time period, using IETC-RFC-3339 format"
      startDateTime:
        type: "string"
        format: "date-time"
        example: "1985-04-12T23:20:50.52Z"
        description: "Start of the time period, using IETC-RFC-3339 format"
    description: "A period of time, either as a deadline (endDateTime only) a startDateTime\
      \ only, or both"
    example:
      startDateTime: "1985-04-12T23:20:50.52Z"
      endDateTime: "1985-04-12T23:20:50.52Z"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "string"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "string"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"